Use saved tokenizers instead of vocab files
